chore: fix the test structure, and refactored testcases
